Refactor duplicate link update flows into one useHandleLinkChange
